finish corrections
